Separate GUI and conversion code
